complete the character creation
I created the character creation inside of the parser for now. I'm hoping to move it out of there and into internal funcs if at all possible. For now this will do.

-I also solved the problem for now with previous inputs. Two solutions:
I created a global variable called prevInput which gets the input and caches it. Or you can actually look into the inputs[] array and find what you need in there.