create a more generalized and cleaner pressenter system
To solve my issue and to make press enter to continue more easier to work with I have devised a solution:
Essentially call pressEnter(roomId you want to go to) and then in the new room in the onEnter(): put renableInput() and that will reenable input with out giving an extra keydown event.

I will do some more testing but it looks pretty good.