finish hote-revi room/ fix wear command / fix story triggers
The hotel room revisited is now completely done, except for bug testing, adding in small things, such as move counts, but all story and items are in the file. Wear command has been updated so IF item is top and bottom works. Need to think about adding a onWear so that it doesn't get clogged up with specific items that have unique rules.