add non buggy js lobby file, add more story nodes, add blocks
created a new js file to go off of in the lobby revisited as there was some problems with my last file. It had to do with the formatting and so I decided creating a new one was easier. Added lots of work, flushing out the story and getting the rooms to have the correct information and blocks.